Refactored `utils.ts`
