+++ packages/foam-vscode/src/features/hover-provider.ts
@@ -1,6 +1,5 @@
 import { uniqWith } from 'lodash';
 import * as vscode from 'vscode';
-import { getNoteTooltip, mdDocSelector, isSome } from '../utils';
 import { fromVsCodeUri, toVsCodeRange } from '../utils/vsc-utils';
 import {
   ConfigurationMonitor,
@@ -15,6 +14,8 @@ import { OPEN_COMMAND } from './commands/open-resource';
 import { CREATE_NOTE_COMMAND } from './commands/create-note';
 import { commandAsURI } from '../utils/commands';
 import { Location } from '../core/model/location';
+import { getNoteTooltip, mdDocSelector } from '../services/editor';
+import { isSome } from '../core/utils';
 
 export const CONFIG_KEY = 'links.hover.enable';
