+++ packages/foam-vscode/src/features/panels/utils/tree-view-utils.ts
@@ -5,11 +5,11 @@ import { toVsCodeUri } from '../../../utils/vsc-utils';
 import { Range } from '../../../core/model/range';
 import { URI } from '../../../core/model/uri';
 import { FoamWorkspace } from '../../../core/model/workspace';
-import { getNoteTooltip } from '../../../utils';
 import { isSome } from '../../../core/utils';
 import { getBlockFor } from '../../../core/services/markdown-parser';
 import { Connection, FoamGraph } from '../../../core/model/graph';
 import { Logger } from '../../../core/utils/log';
+import { getNoteTooltip } from 'packages/foam-vscode/src/services/editor';
 
 export class BaseTreeItem extends vscode.TreeItem {
   resolveTreeItem(): Promise<vscode.TreeItem> {
