+++ packages/foam-vscode/src/features/preview/wikilink-navigation.ts
@@ -2,7 +2,6 @@
 
 import markdownItRegex from 'markdown-it-regex';
 import * as vscode from 'vscode';
-import { isNone } from '../../utils';
 import { FoamWorkspace } from '../../core/model/workspace';
 import { Logger } from '../../core/utils/log';
 import { toVsCodeUri } from '../../utils/vsc-utils';
@@ -10,6 +9,7 @@ import { MarkdownLink } from '../../core/services/markdown-link';
 import { Range } from '../../core/model/range';
 import { isEmpty } from 'lodash';
 import { toSlug } from '../../utils/slug';
+import { isNone } from '../../core/utils';
 
 export const markdownItWikilinkNavigation = (
   md: markdownit,
