+++ packages/foam-vscode/src/features/tag-completion.ts
@@ -1,7 +1,8 @@
 import * as vscode from 'vscode';
 import { Foam } from '../core/model/foam';
 import { FoamTags } from '../core/model/tags';
-import { isInFrontMatter, isOnYAMLKeywordLine, mdDocSelector } from '../utils';
+import { isInFrontMatter, isOnYAMLKeywordLine } from '../core/utils/md';
+import { mdDocSelector } from '../services/editor';
 
 // this regex is different from HASHTAG_REGEX in that it does not look for a
 // #+character. It uses a negative look-ahead for `# `
